server/requirements.txt
django==2.2.13
djangorestframework==3.9.4
requests==2.21.0
psycopg2-binary
python-env-utils==0.4.1
